Migrated printer related util functions to a printer_utils file
